Backbones: export DN42 default routes from BGP
We currently receive a single route for the entire ChaosVPN space
172.31.0.0/16 from our external peers, which has a higher preference
than the static unreachable route we configure for that range. Thus, if
backbone machines are to export this "default" route to access machines,
we need to allow BGP as a default route source in the export filter.
Otherwise, the 172.31.0.0/16 route will only be exported to access
machines when we *aren't* receiving it from external peers, and the
static unreachable route is preferred.

This has got me thinking that there may be a more elegant way to manage
the backbone vs. access BGP views on backbone machines with multiple
routing tables. I think this would address one remaining issue with the
current arrangement, namely that we will experience internal churn if
the 172.31.0.0/16 route is flapping. But that's a project for a later
time.