Create docker_repo role for Docker APT repository
The new role configures the Docker APT repository without installing
anything. The tasks were copied from the docker role, which now has this
as a dependency.

Separating the repository configuration from the installation of Docker
will be useful for Kubernetes installation, which needs the repository
but then only installs containerd (and to configure it specifically for
Kubernetes).