Switch from BGP confederations to route reflectors
After experimenting with adding a third sub-AS connected directly to
both of the existing sub-ASes, it's become apparent that it causes some
issues, at least with BIRD. It looks like BIRD withdraws a route when it
notices that the NEXT_HOP is the same as its peer's IP address. This
occurs frequently in this topology, and it has left a lot of "invalid
NEXT_HOP" messages in the logs. It also seems to be the cause of some
BGP update storms that I've observed.

Since in practice the network is turning out to be neatly hierarchical,
with a central mesh of backbone hosts and access hosts attached to the
backbones in each site, route reflection now seems to be a better
solution. In this new arrangement, backbone hosts continue to mesh
fully, but they now act as route reflectors for the access hosts in
their site. All told, this should actually result in fewer iBGP sessions
than before. The configuration also seems simpler.

So that templates can enumerate all hosts in the current host's
site---used here to autogenerate the peer lists in the BIRD
configuration---a new "site_group" variable is added. It's defined in
the group_vars for each site group, and it just gives the name of the
group.