Overhaul routing and firewall design
The old design (a single AS with all BGP next-hops in OSPF) was designed
with a single site-local network design in mind: the design used for the
oxc homelab. The two homelab VMs were, from a routing perspective, more
or less the same as the lga backbone VPS: all participated in OSPF and
received all internal routes via BGP (with the lga and oxc backbone
nodes also dealing with external DN42 routes). It was assumed that all
of a site's hosts were connected to the same Ethernet network and could
reach a backbone host directly. Notably, the hypervisor was not on this
network.

Keeping this design precludes experimenting with other network
topologies (e.g. leaf-spine) in the new beh homelab. Furthermore, the
old design is difficult to integrate with cloud-provider networks.
Therefore, we have a new system.

* Routers fall into one of two scopes: backbone or site. The
  "router_role" variable is changed to "router_scope".

* Backbone hosts form a Wireguard mesh. (There is no more "wg_mesh"
  group.) Each participates in OSPF and has an IBGP peering in
  AS4242420703 with every other backbone host, except for those in the
  same site. (This limitation may be addressed in the future.)

* Site (non-backbone) hosts do not participate in the backbone OSPF
  network. They use default routes to reach everything except networks
  in the local site. (This is a change from the previous design, in
  which all internal routes, even from other sites, were sent to routers
  with the "access" role.)

* How routing information is transmitted in a single site depends on the
  site. The router_site_role host variable and a site-role-specific BIRD
  configuration file are used as necessary. (There are currently no
  configuration files for any method, since there are no site networks
  right now. Configurations for a leaf-spine setup for the beh homelab are
  in the works.)

* Backbone routers redistribute local routes into AS4242420703. BGP
  running within a site may use ASNs 4200000000 through 4200000999;
  backbone routers remove these ASNs from a route's AS_PATH before
  sending the route to other backbone routers.

* For simplicity, service advertisement ASNs are now hardwired:
  AS4200000998 for the host and AS4200000999 for the container.

* Since site-scope routers receive routes only for the local site
  network, they will always use local instances of anycasted services if
  they are available. (Otherwise, they'll fall back to default routes.)
  Each local network configuration will need to make sure that the
  backbone routers select local anycast routes are over remote ones.

In addition to this work, this commit includes a refresh of host
nftables rules for forwarding. It's now based on verdict maps. (It
probably would have been best to split this into another commit, but the
two reworks were happening in parallel and were getting tangled up a
bit.) One salient feature: "protected" interfaces, which don't allow
non-established-or-related traffic in, except for traffic coming from
explicitly specified source addresses.