Rework HTTP ingress
It's going to be easier to just write manual Caddyfile templates for
each ingress, so this commit makes that change. The Caddy version is
also upgraded to 2.5.2, and permissions on the Caddyfile are fixed.

The DN42 ingress is currently out of commission until I can get ACME
challenges working with anycast.

The next step is to upgrade to Caddy 2.6. It brings HTTP/3 support and
has some minor breaking changes, so that will be its own commit.