Prefer routes originated near where we receive them
When a route is originated near where we receive it, there's a good
chance that traffic never leaves the region and that the latency to the
originating autonomous system's network edge is low.

In particular, this resolves an interesting issue I encountered today.
All five peers at LGA are themselves peered with BURBLE-AS. Four of the
five are peered with BURBLE-AS's Beauharnois server, but one is only
peered with the (much farther away) Frankfurt server. BIRD selected the
path that ultimately went through Frankfurt. With these changes, BIRD
picks one of the other four peers. Latency to BURBLE-AS anycast services
is greatly reduced, since traffic to BURBLE-AS now reaches BURBLE-AS in
Beauharnois, rather than being served from Frankfurt.

As of this commit, we use the following BGP local preference scheme:

100: local routes
 50: external routes originated and received in the same DN42 region
 40: external routes originated in a DN42 region "near" the region in
     which they were received (currently, for North American regions,
     "near" means "one of the other North American regions")
 10: all other external routes