Set higher local preference when AS path length is <= 1
This ensures that if we're directly connected to an AS, traffic destined
for it is sent directly to it (rather than going through other ASes
first). If all routes have the same local preference, then this happens
by default. However, now that we're setting different local preferences
for different external routes, we need to give these direct routes the
highest local preference (among external routes) to ensure that we keep
the same behavior.