Forward packets only from specified interfaces
The current firewall configuration (inadvertently) does not apply any
filters to packets arriving on interfaces that aren't explictly
specified (through the various *_if Ansible host variables). The fix is
to separate the outbound vmap rule into its own chain, "fwd-stage-2",
and to "goto" this chain after a packet succesfully passes any inbound
rules. Then the "forward" chain is a single vmap rule. Thus, any packets
coming in on interfaces not present in the vmap reach the end of the
"forward" chain and are dropped (per the chain policy). In fact, since
we now "goto" the "fwd-from-*" chains and then "goto fwd-stage-2"
---rather than "jump" to the "fwd-from-*" chains and then "return" to
"forward"---packets that are not matched by any rule in the "fwd-from-*"
chains will be dropped (by "forward" chain policy), rather than returned
to "forward" as they were previously. Therefore, we no longer need the
catch-all "drop" rules at the end of the "fwd-from-*" chains. This is
much nicer!